+++ src/main/java/gov/nasa/pds/harvest/meta/AutogenExtractor.java
@@ -12,7 +12,7 @@ import org.w3c.dom.NodeList;
 
 import gov.nasa.pds.harvest.Constants;
 import gov.nasa.pds.harvest.cfg.model.AutogenCfg;
-import gov.nasa.pds.harvest.util.FieldMap;
+import gov.nasa.pds.registry.common.util.FieldMap;
 import gov.nasa.pds.registry.common.util.date.PdsDateConverter;
 import gov.nasa.pds.registry.common.util.xml.NsUtils;
 import gov.nasa.pds.registry.common.util.xml.XmlDomUtils;
