+++ src/main/java/gov/nasa/pds/harvest/meta/InternalReferenceExtractor.java
@@ -7,7 +7,7 @@ import org.w3c.dom.Document;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
-import gov.nasa.pds.harvest.util.FieldMapSet;
+import gov.nasa.pds.registry.common.util.FieldMapSet;
 import gov.nasa.pds.registry.common.util.xml.XPathUtils;
