+++ src/main/java/gov/nasa/pds/harvest/meta/Metadata.java
@@ -2,8 +2,8 @@ package gov.nasa.pds.harvest.meta;
 
 import java.util.Set;
 
-import gov.nasa.pds.harvest.util.FieldMapList;
-import gov.nasa.pds.harvest.util.FieldMapSet;
+import gov.nasa.pds.registry.common.util.FieldMapList;
+import gov.nasa.pds.registry.common.util.FieldMapSet;
 
 
 /**
